requirements.txt
groq
notion-client==2.2.1
python-dotenv
mcp